reorganize modules and minor tweaks
