+++ pliers/stimuli/video.py
@@ -11,6 +11,7 @@ from .image import ImageStim
 class VideoFrameStim(ImageStim):
 
     ''' A single frame of video.
+
     Args:
         video (VideoStim): The source VideoStim the frame is drawn from.
         frame_num (int): The index of the current frame in the source video.
@@ -37,6 +38,7 @@ class VideoFrameStim(ImageStim):
 class VideoFrameCollectionStim(Stim):
 
     ''' A collection of video frames.
+
     Args:
         filename (str): Path to input file, if one exists.
         frame_index (list): List of indices of frames retained from the
@@ -83,6 +85,7 @@ class VideoFrameCollectionStim(Stim):
 
     def get_frame(self, index=None, onset=None):
         ''' Get video frame at the specified index.
+
         Args:
             index (int): Positional index of the desired frame.
             onset (float): Onset (in seconds) of the desired frame.
@@ -116,8 +119,10 @@ class VideoFrameCollectionStim(Stim):
 
     def save(self, path):
         ''' Save source video to file.
+
         Args:
             path (str): Filename to save to.
+
         Notes: Saves entire source video to file, not just currently selected
             frames.
         '''
@@ -128,6 +133,7 @@ class VideoFrameCollectionStim(Stim):
 class VideoStim(VideoFrameCollectionStim):
 
     ''' A video.
+
     Args:
         filename (str): Path to input file, if one exists.
         onset (float): Optional onset of the video file (in seconds) with
