+++ pliers/support/__init__.py
@@ -1,1 +1,9 @@
-__all__ = ['decorators', 'exceptions']
+''' Support utilities. '''
+
+from .decorators import requires_nltk_corpus
+from .download import download_nltk_data
+
+__all__ = [
+    'requires_nltk_corpus',
+    'download_nltk_data'
+]
