+++ pliers/tests/test_io.py
@@ -3,12 +3,9 @@ from pliers.stimuli import (load_stims, AudioStim, ImageStim)
 from pliers.extractors import (STFTAudioExtractor,
                                GoogleVisionAPIFaceExtractor,
                                ExtractorResult)
-from pliers.export import to_long_format
 from pliers.graph import Graph
 from os.path import join
 from six import string_types
-import pytest
-
 import pytest
 
 
@@ -51,24 +48,6 @@ def test_loader_nonexistent():
 
     with pytest.raises(IOError):
         stims = load_stims(audio_file, fail_silently=True)
-
-
-def test_convert_to_long():
-    audio_dir = join(get_test_data_path(), 'audio')
-    stim = AudioStim(join(audio_dir, 'barber.wav'))
-    ext = STFTAudioExtractor(frame_size=1., spectrogram=False,
-                        freq_bins=[(100, 300), (300, 3000), (3000, 20000)])
-    timeline = ext.transform(stim)
-    long_timeline = to_long_format(timeline)
-    assert long_timeline.shape == (timeline.to_df().shape[0] * 3, 4)
-    assert 'feature' in long_timeline.columns
-    assert 'value' in long_timeline.columns
-    assert '100_300' not in long_timeline.columns
-    timeline = ExtractorResult.merge_features([timeline])
-    long_timeline = to_long_format(timeline)
-    assert 'feature' in long_timeline.columns
-    assert 'extractor' in long_timeline.columns
-    assert '100_300' not in long_timeline.columns
 
 
 @pytest.mark.skipif("'GOOGLE_APPLICATION_CREDENTIALS' not in os.environ")
