+++ pliers/tests/test_transformers.py
@@ -79,10 +79,10 @@ def test_validation_levels(caplog):
     res = ext.transform(stim, validation='warn')
 
     log_message = caplog.records[0].message
-    assert log_message == "Transformers of type BrightnessExtractor can "\
-                  "only be applied to stimuli  of type(s) <class 'pliers"\
-                  ".stimuli.image.ImageStim'> (not type TextStim), and no "\
-                  "applicable Converter was found."
+    assert log_message == ("Transformers of type BrightnessExtractor can "
+                  "only be applied to stimuli of type(s) <class 'pliers"
+                  ".stimuli.image.ImageStim'> (not type TextStim), and no "
+                  "applicable Converter was found.")
     assert not res
 
     res = ext.transform(stim, validation='loose')
