+++ pliers/utils/__init__.py
@@ -4,6 +4,7 @@ from .base import (listify, flatten, batch_iterable, classproperty, isiterable,
                    isgenerator, progress_bar_wrapper, attempt_to_import,
                    EnvironmentKeyMixin, verify_dependencies)
 
+
 __all__ = [
     'listify',
     'flatten',
